backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
xai-sdk>=0.1.0
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.27.0
